feat(tokens): Adding `outline-tokens`, lots of basic refactoring.
