+++ packages/outline-config/tailwind.config.js
@@ -104,78 +104,6 @@ module.exports = {
         800: 'var(--outline-blue-800)',
         900: 'var(--outline-blue-900)',
       },
-      'red': {
-        50: 'var(--red-50)',
-        100: 'var(--red-100)',
-        200: 'var(--red-200)',
-        300: 'var(--red-300)',
-        400: 'var(--red-400)',
-        500: 'var(--red-500)',
-        600: 'var(--red-600)',
-        700: 'var(--red-700)',
-        800: 'var(--red-800)',
-        900: 'var(--red-900)',
-      },
-      'yellow': {
-        50: 'var(--yellow-50)',
-        100: 'var(--yellow-100)',
-        200: 'var(--yellow-200)',
-        300: 'var(--yellow-300)',
-        400: 'var(--yellow-400)',
-        500: 'var(--yellow-500)',
-        600: 'var(--yellow-600)',
-        700: 'var(--yellow-700)',
-        800: 'var(--yellow-800)',
-        900: 'var(--yellow-900)',
-      },
-      'green': {
-        50: 'var(--green-50)',
-        100: 'var(--green-100)',
-        200: 'var(--green-200)',
-        300: 'var(--green-300)',
-        400: 'var(--green-400)',
-        500: 'var(--green-500)',
-        600: 'var(--green-600)',
-        700: 'var(--green-700)',
-        800: 'var(--green-800)',
-        900: 'var(--green-900)',
-      },
-      'indigo': {
-        50: 'var(--red-50)',
-        100: 'var(--red-100)',
-        200: 'var(--red-200)',
-        300: 'var(--red-300)',
-        400: 'var(--red-400)',
-        500: 'var(--red-500)',
-        600: 'var(--red-600)',
-        700: 'var(--red-700)',
-        800: 'var(--red-800)',
-        900: 'var(--red-900)',
-      },
-      'purple': {
-        50: 'var(--purple-50)',
-        100: 'var(--purple-100)',
-        200: 'var(--purple-200)',
-        300: 'var(--purple-300)',
-        400: 'var(--purple-400)',
-        500: 'var(--purple-500)',
-        600: 'var(--purple-600)',
-        700: 'var(--purple-700)',
-        800: 'var(--purple-800)',
-        900: 'var(--purple-900)',
-      },
-      'pink': {
-        50: 'var(--pink-50)',
-        100: 'var(--pink-100)',
-        200: 'var(--pink-200)',
-        300: 'var(--pink-300)',
-        400: 'var(--pink-400)',
-        500: 'var(--pink-500)',
-        600: 'var(--pink-600)',
-        700: 'var(--pink-700)',
-        800: 'var(--pink-800)',
-        900: 'var(--pink-900)',
-      },
     },
     fontSize: {
       'xs': ['var(--fs-xs)', { lineHeight: 'var(--lh-xs)' }],
