+++ packages/outline-storybook/package.json
@@ -12,7 +12,7 @@
     "directory": "packages/outline-storybook"
   },
   "dependencies": {
-    "@phase2/outline-demonstrations": "^0.0.1",
+    "@phase2/outline-style-guide": "^0.0.1",
     "lit": "^2.0.2"
   },
   "publishConfig": {
