+++ packages/outline-storybook/stories/components/demonstration/outline-swatch.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Canvas } from '@storybook/addon-docs';
 import { html } from 'lit';
-import { OutlineSwatch, swatchSets } from '@phase2/outline-demonstrations';
+import { OutlineSwatch, swatchSets } from '@phase2/outline-style-guide';
 
 <Meta
   title="Utilities/Swatch"
