+++ packages/outline-storybook/stories/guides/welcome.stories.mdx
@@ -1,7 +1,7 @@
 import { html } from 'lit';
 import { Meta, Canvas, Story, Preview } from '@storybook/addon-docs';
 
-import '@phase2/outline-demonstrations';
+import '@phase2/outline-style-guide';
 
 <!-- Importing the needed components for default stories -->
 import { OutlineContainer } from '@phase2/outline-container';
