+++ packages/outline-storybook/stories/tokens/colors.stories.mdx
@@ -1,6 +1,6 @@
 import { html, unsafeHTML } from 'lit';
 import { Meta, Canvas, Story, Preview } from '@storybook/addon-docs';
-import '@phase2/outline-demonstrations'
+import '@phase2/outline-style-guide'
 
 <Meta
   title="Design Tokens/Colors/Solid"
