+++ packages/outline-style-guide/src/outline-swatch/outline-swatch.css
